requirements.txt
# Core NLP
spacy==3.8.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# Speech to Text
openai-whisper==20250625
torch==2.8.0
tiktoken==0.12.0

# Video & Audio Processing
moviepy==2.2.1
opencv-python==4.12.0.88
imageio==2.37.2
imageio-ffmpeg==0.6.0

# Numerical & Utilities
numpy==2.0.2
tqdm==4.67.1
requests==2.32.5
python-dotenv==1.2.1

# NOTE:
# FFmpeg is required but must be installed separately (system dependency)
